feat(App-DeleteUserCommandHandler): unify the deletes
